requirements.txt
torch==2.1.0
torchvision==0.16.0
scikit-learn==1.3.2
numpy==1.26.2
joblib==1.3.2
Pillow==10.1.0
gradio==4.8.0
tensorflow==2.15.0
pandas
seaborn